# docs/swagger.yaml
basePath: /api
definitions:
  handler.AddToCartRequest:
    properties:
      hb_after:
        example: 120
        type: integer
      hb_before:
        example: 140
        type: integer
      patient_height:
        example: 1.75
        type: number
      patient_weight:
        example: 70
        type: integer
      surgery_duration:
        example: 2.5
        type: number
      total_blood_loss:
        example: 300
        type: integer
    type: object
  handler.AddToCartResponse:
    properties:
      bloodlosscalc:
        properties:
          id:
            example: 5
            type: integer
          patient_height:
            example: 1.75
            type: number
          patient_weight:
            example: 70
            type: integer
          status:
            example: черновик
            type: string
        type: object
      is_new_request:
        example: true
        type: boolean
      message:
        example: Operation added to bloodlosscalc successfully
        type: string
      operation:
        properties:
          avg_blood_loss:
            example: 150
            type: integer
          blood_loss_coeff:
            example: 0.04
            type: number
          id:
            example: 1
            type: integer
          image_url:
            type: string
          title:
            example: Аппендэктомия
            type: string
        type: object
      service_count:
        example: 1
        type: integer
    type: object
  handler.AuthRequest:
    properties:
      password:
        example: admin123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handler.AuthResponse:
    properties:
      expires_at:
        example: "2024-12-11T14:30:00Z"
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/handler.UserResponse'
    type: object
  handler.BLItem:
    properties:
      avg_blood_loss:
        example: 150
        type: integer
      blood_loss_coeff:
        example: 0.04
        type: number
      hb_after:
        example: 120
        type: integer
      hb_before:
        example: 140
        type: integer
      operation_image:
        type: string
      operation_title:
        example: Аппендэктомия
        type: string
      surgery_duration:
        example: 2.5
        type: number
      total_blood_loss:
        example: 300
        type: integer
    type: object
  handler.BloodlosscalcDetailResponse:
    properties:
      calculated_count:
        type: integer
      completed_at:
        type: string
      created_at:
        example: 11.12.2024
        type: string
      creator:
        example: admin
        type: string
      formed_at:
        type: string
      id:
        example: 5
        type: integer
      items:
        items:
          $ref: '#/definitions/handler.BLItem'
        type: array
      moderator:
        type: string
      patient_height:
        example: 1.75
        type: number
      patient_weight:
        example: 70
        type: integer
      status:
        example: черновик
        type: string
    type: object
  handler.BloodlosscalcResponse:
    properties:
      calculated_count:
        type: integer
      completed_at:
        type: string
      created_at:
        example: 11.12.2024
        type: string
      creator_login:
        example: admin
        type: string
      formed_at:
        type: string
      id:
        example: 5
        type: integer
      moderator_login:
        type: string
      patient_height:
        example: 1.75
        type: number
      patient_weight:
        example: 70
        type: integer
      status:
        example: черновик
        type: string
    type: object
  handler.CartInfoResponse:
    properties:
      current_request_id:
        example: 5
        type: integer
      service_count:
        example: 3
        type: integer
    type: object
  handler.ErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      description:
        example: Ошибка авторизации
        type: string
      status:
        example: error
        type: string
    type: object
  handler.MessageResponse:
    description: Стандартный ответ с сообщением
    properties:
      message:
        example: Операция выполнена успешно
        type: string
    type: object
  handler.Operation:
    properties:
      avg_blood_loss:
        example: 150
        type: integer
      blood_loss_coeff:
        example: 0.04
        type: number
      description:
        example: Удаление червеобразного отростка
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: http://localhost:9000/blood-loss-images/appendectomy.jpg
        type: string
      status:
        example: активна
        type: string
      title:
        example: Аппендэктомия
        type: string
    type: object
  handler.OperationsListResponse:
    properties:
      operations:
        items:
          $ref: '#/definitions/handler.Operation'
        type: array
    type: object
  handler.RegisterRequest:
    properties:
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: ivanov
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  handler.RegisterResponse:
    properties:
      message:
        example: Пользователь успешно зарегистрирован
        type: string
      user:
        properties:
          is_moderator:
            example: false
            type: boolean
          user_id:
            example: 1
            type: integer
          username:
            example: ivanov
            type: string
        type: object
    type: object
  handler.UpdateBloodlosscalcRequest:
    properties:
      patient_height:
        example: 1.8
        type: number
      patient_weight:
        example: 75
        type: integer
    type: object
  handler.UpdateOperationRequest:
    properties:
      hb_after:
        example: 110
        type: integer
      hb_before:
        example: 130
        type: integer
      surgery_duration:
        example: 3
        type: number
      total_blood_loss:
        example: 350
        type: integer
    type: object
  handler.UpdateUserRequest:
    properties:
      password:
        example: new_password123
        type: string
      username:
        example: new_username
        type: string
    type: object
  handler.UserResponse:
    properties:
      is_moderator:
        example: true
        type: boolean
      user_id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Стандартный успешный ответ
  title: Blood Loss Calculator API
  version: 1.0.0
paths:
  /api/auth:
    post:
      consumes:
      - application/json
      description: Вход в систему с получением JWT токена
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Аутентификация
  /api/bloodlosscalc_operations:
    delete:
      consumes:
      - application/json
      description: Удаление операции из заявки (только для черновиков)
      parameters:
      - description: ID заявки
        in: query
        name: bloodlosscalc_id
        required: true
        type: integer
      - description: ID операции
        in: query
        name: operation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Не автор заявки или не черновик
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить операцию из заявки
      tags:
      - Операции в заявке
    put:
      consumes:
      - application/json
      description: Обновление данных операции (гемоглобин, длительность и т.д.)
      parameters:
      - description: ID заявки
        in: query
        name: bloodlosscalc_id
        required: true
        type: integer
      - description: ID операции
        in: query
        name: operation_id
        required: true
        type: integer
      - description: Новые данные операции
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Не автор заявки или не черновик
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить операцию в заявке
      tags:
      - Операции в заявке
  /api/bloodlosscalcs:
    get:
      description: Получение списка заявок текущего пользователя с фильтрацией
      parameters:
      - description: Фильтр по статусу
        enum:
        - черновик
        - сформирована
        - завершена
        - удален
        in: query
        name: status
        type: string
      - description: Дата от (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Дата до (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.BloodlosscalcResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить заявки пользователя
      tags:
      - Заявки
  /api/bloodlosscalcs/{id}:
    delete:
      consumes:
      - application/json
      description: Логическое удаление заявки (статус → "удален")
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Не автор заявки
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить заявку
      tags:
      - Заявки
    get:
      description: Получение детальной информации о заявке со списком операций
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BloodlosscalcDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа к заявке
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить заявку по ID
      tags:
      - Заявки
    put:
      consumes:
      - application/json
      description: Изменение роста и веса пациента в заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные пациента
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateBloodlosscalcRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Не автор заявки
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить данные заявки
      tags:
      - Заявки
  /api/bloodlosscalcs/{id}/form:
    put:
      consumes:
      - application/json
      description: Перевод заявки из статуса "черновик" в "сформирована"
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Не все данные заполнены
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Не автор заявки
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Сформировать заявку
      tags:
      - Заявки
  /api/logout:
    post:
      consumes:
      - application/json
      description: Завершение сессии пользователя, токен добавляется в blacklist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - Аутентификация
  /api/operationcart:
    get:
      description: Получение ID текущей заявки и количества услуг в ней
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CartInfoResponse'
      summary: Получить информацию о корзине
      tags:
      - Корзина
  /api/operations:
    get:
      description: Получение списка операций с возможностью фильтрации
      parameters:
      - description: Фильтр по названию операции
        in: query
        name: title
        type: string
      - description: Фильтр по статусу
        enum:
        - активна
        - неактивна
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OperationsListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить список операций
      tags:
      - Операции
  /api/operations/{id}:
    get:
      description: Получение детальной информации об операции
      parameters:
      - description: ID операции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Operation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить операцию по ID
      tags:
      - Операции
  /api/operations/{id}/add_to_bloodlosscalc:
    post:
      consumes:
      - application/json
      description: Добавление операции в текущую заявку пользователя (корзину)
      parameters:
      - description: ID операции
        in: path
        name: id
        required: true
        type: integer
      - description: Дополнительные данные (рост, вес, показатели крови)
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AddToCartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Операция уже в заявке
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить операцию в заявку
      tags:
      - Операции
  /api/register:
    post:
      consumes:
      - application/json
      description: Создание учетной записи пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Аутентификация
  /api/user:
    get:
      description: Получение данных текущего авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить профиль пользователя
      tags:
      - Пользователь
    put:
      consumes:
      - application/json
      description: Изменение данных пользователя (имя, пароль)
      parameters:
      - description: Новые данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Имя пользователя уже занято
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль пользователя
      tags:
      - Пользователь
securityDefinitions:
  BearerAuth:
    description: JWT токен в формате "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"